refactoring(15): extract renderReviewScores from ResultsPage
Move the review-scores render block (the third low-risk pure render
block) into resultSections.tsx as an exported render-helper function.
Its three container-local helpers (decisionVariant/decisionLabel/
renderDimension), used only here, are nested inside verbatim so every
call site is unchanged and t is not threaded. ResultsPage 1744->1446.
